requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
twilio>=8.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
